refactor: modularization (#1)
* refactor: initial modularization of the codebase

Split monolithic lib.rs into a more maintainable structure:
- Create games/ directory for game implementations
- Move Perception (Seek) game into games/perception.rs
- Create utils.rs for shared functionality
- Update lib.rs to re-export necessary items

This refactoring sets up the foundation for:
1. Adding remaining cognitive test modules
2. Better separation of concerns
3. Easier maintenance and testing
4. More idiomatic Rust module structure

* break down the Perception module further
Game state management (saving/loading/creation)
Rendering logic
Movement and maze interaction
Core game mechanics (walls, positions, etc.)

* fix compilation

* extracting the timer-related functionality into a new module

* Create a new input.rs module for handling user interactions

* Removed redundant game.render()? call since it's already done in new()
